Backend Tests Updated (#468)
* fix some tests

* Add EnrolleeProfileVersion Faker

* fix all tests except enrolment certificate tests

* fix org type loading

* Removed int cast

+++ prime-dotnet-webapi-tests/Mocks/BaseMockService.cs
@@ -9,9 +9,9 @@ namespace PrimeTests.Mocks
     {
         public const int DEFAULT_ENROLMENTS_SIZE = 5;
         public const int DEFAULT_ENROLLEES_SIZE = 5;
-        protected static short NULL_STATUS_CODE = -1;
+        protected static int NULL_STATUS_CODE = -1;
 
-        private static Dictionary<short, Status> _statusMap = new Dictionary<short, Status> {
+        private static Dictionary<int, Status> _statusMap = new Dictionary<int, Status> {
             { NULL_STATUS_CODE, new Status { Code = NULL_STATUS_CODE, Name = "No Status" } },
             { Status.ACTIVE_CODE, new Status { Code = Status.ACTIVE_CODE, Name = "Active" } },
             { Status.UNDER_REVIEW_CODE, new Status { Code = Status.UNDER_REVIEW_CODE, Name = "Under Review" } },
@@ -87,14 +87,14 @@ namespace PrimeTests.Mocks
             _fakeDb.Add(ENROLLEE_KEY, new Dictionary<int, Enrollee>());
 
             // seed the lookup tables
-            _fakeDb.Add(COLLEGE_KEY, new Dictionary<short, College> {
+            _fakeDb.Add(COLLEGE_KEY, new Dictionary<int, College> {
                 { 1, new College { Code = 1, Name = "College of Physicians and Surgeons of BC (CPSBC)", Prefix = "91" } },
                 { 2, new College { Code = 2, Name = "College of Pharmacists of BC (CPBC)", Prefix = "P1" } },
                 { 3, new College { Code = 3, Name = "College of Registered Nurses of BC (CRNBC)", Prefix = "96" } },
                 { 4, new College { Code = 4, Name = "None", Prefix = null } }
             });
 
-            _fakeDb.Add(LICENSE_KEY, new Dictionary<short, License> {
+            _fakeDb.Add(LICENSE_KEY, new Dictionary<int, License> {
                 { 1, new License { Code = 1, Name = "Full - General" } },
                 { 2, new License { Code = 2, Name = "Full - Pharmacist" } },
                 { 3, new License { Code = 3, Name = "Full - Specialty" } },
@@ -102,32 +102,32 @@ namespace PrimeTests.Mocks
                 { 5, new License { Code = 5, Name = "Temporary Registered Nurse" } }
             });
 
-            _fakeDb.Add(PRACTICE_KEY, new Dictionary<short, Practice> {
+            _fakeDb.Add(PRACTICE_KEY, new Dictionary<int, Practice> {
                 { 1, new Practice { Code = 1, Name = "Remote Practice" } },
                 { 2, new Practice { Code = 2, Name = "Reproductive Care" } },
                 { 3, new Practice { Code = 3, Name = "Sexually Transmitted Infections (STI)" } },
                 { 4, new Practice { Code = 4, Name = "None" } }
             });
 
-            _fakeDb.Add(JOB_NAME_KEY, new Dictionary<short, JobName> {
+            _fakeDb.Add(JOB_NAME_KEY, new Dictionary<int, JobName> {
                 { 1, new JobName { Code = 1, Name = "Medical Office Assistant" } },
                 { 4, new JobName { Code = 2, Name = "Pharmacy Assistant" } },
                 { 6, new JobName { Code = 3, Name = "Registration Clerk" } },
                 { 7, new JobName { Code = 4, Name = "Ward Clerk" } }
             });
 
-            _fakeDb.Add(ORGANIZATION_TYPE_KEY, new Dictionary<short, OrganizationType> {
+            _fakeDb.Add(ORGANIZATION_TYPE_KEY, new Dictionary<int, OrganizationType> {
                 { 1, new OrganizationType { Code = 1, Name = "Health Authority" } },
                 { 2, new OrganizationType { Code = 2, Name = "Pharmacy" } }
             });
 
             _fakeDb.Add(STATUS_KEY, _statusMap);
 
-            _fakeDb.Add(COUNTRY_KEY, new Dictionary<short, Country> {
+            _fakeDb.Add(COUNTRY_KEY, new Dictionary<int, Country> {
                 { 1, new Country { Code = "CA", Name = "Canada" } }
             });
 
-            _fakeDb.Add(PROVINCE_KEY, new Dictionary<short, Province> {
+            _fakeDb.Add(PROVINCE_KEY, new Dictionary<int, Province> {
                 { 1, new Province { Code = "AB", Name = "Alberta" } },
                 { 2, new Province { Code = "BC", Name = "British Columbia" } },
                 { 3, new Province { Code = "MB", Name = "Manitoba" } },
@@ -143,7 +143,7 @@ namespace PrimeTests.Mocks
                 { 13, new Province { Code = "YT", Name = "Yukon" } }
             });
 
-            _fakeDb.Add(STATUS_REASON_KEY, new Dictionary<short, StatusReason> {
+            _fakeDb.Add(STATUS_REASON_KEY, new Dictionary<int, StatusReason> {
                 { 1, new StatusReason { Code = 1, Name = "Automatic" } },
                 { 2, new StatusReason { Code = 2, Name = "Manual" } },
                 { 3, new StatusReason { Code = 2, Name = "Name Discrepancy" } },
@@ -154,11 +154,11 @@ namespace PrimeTests.Mocks
                 { 8, new StatusReason { Code = 2, Name = "Contact address or Identity Address Out of British Columbia" } }
             });
 
-            _fakeDb.Add(PRIVILEGE_GROUP_KEY, new Dictionary<short, PrivilegeGroup> {
+            _fakeDb.Add(PRIVILEGE_GROUP_KEY, new Dictionary<int, PrivilegeGroup> {
                 { 1, new PrivilegeGroup { Code = 4, PrivilegeTypeCode = 1, Name = "Roles" } }
             });
 
-            _fakeDb.Add(PRIVILEGE_TYPE_KEY, new Dictionary<short, PrivilegeType> {
+            _fakeDb.Add(PRIVILEGE_TYPE_KEY, new Dictionary<int, PrivilegeType> {
                 { 1, new PrivilegeType { Code = 1, Name = "Role" } }
             });
